UtilityMethodTestCase: add `Before/After[Class]` attributes
PHPUnit 10 introduced attributes as replacements for docblock annotations.
PHPUnit 11 deprecates the use of docblock annotations in favour of attributes.

If both an attribute as well as an annotation are found, no PHPUnit deprecation warning will be thrown.

This commit adds the `Before/After*` attributes in the user-facing `UtilityMethodTestCase` class to prevent such deprecation notices.

_The "annotations to attributes" conversion for the PHPCSUtils native test suite is not yet necessary and is left for the update to PHPUnit 12._

The `@before/after*` annotations remain as the tests also still need to run on PHP 5.4 - 8.0 using PHPUnit 4.x - 9.x.
These can be removed once the codebase has a PHP 8.1/PHPUnit 10 minimum requirement.

Note: due to the syntax for attributes, these can be safely added as they are ignored as comments on PHP < 8.0.
Along the same line, if there is no "listener" for the attributes (PHP 8.0/PHPUnit 9.x), they are ignored by PHP as well.

Also note that the addition of the attributes, now adds an _explicit_ runtime dependency on PHPUnit for the `UtilityMethodTestCase`, while previously this dependency was _implicit_.
By rights, this means that the `phpunit/phpunit` package dependency should be moved in the `composer.json` file from `require-dev` to `require`. I deem this undesirable though, as for most _end-users_, who use external standards which use PHPCSUtils, it won't be a dependency. The dependency is realistically only for the _tests_ of external standards using the `UtilityMethodTestCase`, in which case, the external standard is expected to have a dependency on PHPUnit anyway.