Tests/AbstractArrayDeclarationSniffTest: work around for `TestCase::onConsecutiveCalls()` deprecation
The `TestCase::onConsecutiveCalls()` is soft deprecated in PHPUnit 10, hard deprecated in PHPUnit 11 and will be removed in PHPUnit 12.

While it is not strictly necessary to address this deprecation now, I've chosen to do so anyway.

Note: the method call is not directly changed from `->will($this->onConsecutiveCalls(...))` to `->willReturn(...)` - which is the PHPUnit native recommendation -, as this would make the test incompatible with older PHPUnit versions.

Additionally, a previous fix to work around the deprecation of `InvocationMocker->withConsecutive()` already uses the `willReturnCallback()` method, so the return value expectations will now need to be set via the `ExpectWithConsecutiveArgs::setExpectationWithConsecutiveArgs()` method.

Refs:
* sebastianbergmann/phpunit 5423
* sebastianbergmann/phpunit 5424