GH Actions/Basics: use reusable workflow
A number of reusable workflows have been introduced in the `PHPCSStandards/.github` repository for workflows used in multiple repos in this organisation, for which the steps are basically the same everywhere.

This will make maintenance of these workflows more straight-forward.

This commit switches a number of the basic QA workflows over to start using the reusable workflows.