UtilityMethodTestCase: make compatible with PHPCS 4.0
This commit introduces a test double for the PHPCS native `Ruleset` class.

PHPCS >= 4.0 is stricter about registering sniffs, so when "mocking" a registered sniff, it will throw a "No sniffs were registered." `RuntimeException`.

Most utility method tests don't need for any sniffs to be registered, they just need a `Ruleset` object to be available to allow for creating a `File` object. This test double accommodates that by catching the exception.