BCFile::getDeclarationName(): remove polyfill
Follow up on 653

PR 644 (re-)introduced polyfill code for the `BCFile::getDeclarationName()` method to account for a change to the method made in PHPCS 3.12.0 without updating the minimum supported PHPCS version.

As the minimum supported PHPCS version has now been updated to 3.13.0 via 653, this polyfill code can be removed again.