WIP | Drop support for PHPUnit < 8 / NEEDS VERY CRITICAL REVIEW
Note: this doesn't need a new major - even though the supported functionality for the UtilityMethodTestCase has changed - as the minimum PHP version bump should ensure that the TestCase is only used with PHPUnit 8 or above.

TODO: look into PHPStan - setting the `phpVersion` results in a shitload of "undefined method" errors related to the tests.

NOTE: method rename of fixture methods can only be done in a major. Then again, leaving these methods as-is shouldn't do any harm.