WAIT - Utils 2.0 | ObjectDeclarations::getName(): stop accepting tokens for non-named structures
The `ObjectDeclarations::getName()` method - for historic reasons - accepted the `T_CLOSURE` and `T_ANON_CLASS` tokens, even though these structures will never have a name, and returned `null` for those tokens.

This commit changes the `ObjectDeclarations::getName()` method to no longer accept those tokens and throw an exception if they are passed to the method instead.

As a secondary change, when the name of a valid structure cannot be determined, the method will now no longer return `null`, but will return an empty string.
This normalizes the return type of the method to always return a string (or throw an exception).

Includes updated unit tests to match.

This change mirrors the upstream change made to the `File::getDeclarationName()` method in PHPCS 4.0.

Ref: PHPCSStandards/PHP_CodeSniffer 1007

Fixes 675