Parentheses: remove compat code for PHPCS 3.x
A number of language constructs, like closure `use` and `isset`, were already handled as parenthesis owners by the `Parentheses` class, prior to PHPCS adding support for recognizing these as parenthesis owners in PHPCS 4.0.0.

As support for PHPCS < 4.0 is being dropped, the special handling of these tokens is no longer needed.