UseStatements::getType(): bug fix for unfinished closures
Of the new tests, test case 4 and 5, would previously both return `"import"` as the use statement type determination. This is clearly wrong.
Additionally, for test case 3, it can also still be reliably determined that this is a closure `use`, while it would previously return an empty string as indicator of "undetermined".

So when there is a parse error, but it can still be reliably determined that the `use` keyword is used for a closure `use` statement, the `UseStatements::getType()` method will now return `"closure"`.

Includes tests.