BCFile::findExtendedClassName(): sync with PHPCS 4.0 / handling of namespace relative parent classes
The PHPCS `File::findExtendedClassName()` method did not handle namespace relative parent classes correctly prior to PHPCS 4.0.

The PHPCSUtils native `ObjectDeclarations::findExtendedClassName()` method **_did_** already handle this correctly.

This commit adds the PHPCS 4.x version of the `findExtendedClassName()` method to the `BCFile` class.

Includes moving related tests from the "Diff" test file to the "normal" test file and updating the docs to annotate this is no longer a difference between the PHPCS native and PHPCSUtils versions of the method.