GetVersionTest: update to allow for PHPCS 4.x releases
This commit updates both the `GetVersionTest` class, as well as the GH Actions script which auto-updates the script, to allow for the `4.x-dev` branch.

It also removes the work-around which was in place for the experimental job against `4.x-dev` in the GH Actions test script. This should no longer be needed.