TODO: update version nr in @since tags | Operators: use class constant for constant array
This commit replaces the (`private`) `Operators::$extraUnaryIndicators` property with the `Operators::UNARY_INDICATORS` class constant, which also includes nearly all other token arrays which need to be checked.
This should give a tiny performance improvement.

This array should never be changed at runtime and doing so can break (other) sniffs, which is why this should always have been a constant.
However, that was previously not possible as constant arrays were not supported prior to PHP 5.6.