TODO: update version nr in @since tags | Lists: use class constant for constant array
This commit replaces the (`private`) `Lists::$listItemDefaults` property with the `Lists::LIST_ITEM_DEFAULTS` class constant.

This array should never be changed at runtime and doing so can break (other) sniffs, which is why this should always have been a constant.
However, that was previously not possible as constant arrays were not supported prior to PHP 5.6.