TODO: update version nr in @since tags | FunctionDeclarations: use class constant for constant array
This commit deprecates the `FunctionDeclarations::$magicFunctions`, `FunctionDeclarations::$magicMethods` and `FunctionDeclarations::$methodsDoubleUnderscore` properties in favour of respectively the `FunctionDeclarations::MAGIC_FUNCTIONS`, `FunctionDeclarations::MAGIC_METHODS` and `FunctionDeclarations::METHODS_DOUBLE_UNDERSCORE` class constants.

This array should never be changed at runtime and doing so can break (other) sniffs, which is why this should always have been a constant.
However, that was previously not possible as constant arrays were not supported prior to PHP 5.6.