TODO: update version nr in @since tags | TypeString: use class constant for constant array
This commit replaces the (`private`) `TypeString::$keywordTypes` property with the `TypeString::KEYWORD_TYPES` class constant.

This array should never be changed at runtime and doing so can break (other) sniffs, which is why this should always have been a constant.
However, that was previously not possible as constant arrays were not supported prior to PHP 5.6.