AbstractArrayDeclarationSniff::getActualArrayKey(): minor simplification
The tokenization of fully qualified `true`/`false`/`null` has been changed for both PHPCS 3.x (as of 3.13.3) as well as PHPCS 4.0.

The new tokenization means we don't need to take these tokens being tokenized as `T_STRING` or `T_NAME_FULLY_QUALIFIED` into account anymore, as long as the minimum supported PHPCS version will be PHPCS 3.13.3 (which it will be since the merge of PR 698).

This commit removes the code which has now become redundant.

Ref:
* PHPCSStandards/PHP_CodeSniffer 1201
* PHPCSStandards/PHP_CodeSniffer 1206