:sparkles: New `PHPCSUtils\Utils\AttributeBlock::appliesTo()` method
This commit adds an additional method to the `AttributeBlock` utility class to allow for finding the language construct (function, constant, parameter etc) an arbitrary attribute applies to.

This method can be passed any token in an attribute block (opener, closer or any token within) and will try to determine the target of the attribute.

The return value will be the stack pointer to a `T_CONST`, `T_FUNCTION`, `T_CLOSURE`, `T_FN`, `T_CLASS`, `T_ANON_CLASS`, `T_INTERFACE`, `T_TRAIT`, `T_ENUM` or `T_VARIABLE` token; or `false` if the attribute target could not be determined.

For multi-property declarations, the `T_VARIABLE` token returned will be to the first property in the declaration.

Mind: this method includes some parse error tolerance for keywords used in an incorrect context. This is deliberate.

Includes extensive unit tests.

Related to 616