Cache Tests: fix PHP 8.5 warning for float cast
PHP 8.5 introduces a new warning about casting floats which are outside the range of supported integers to int.
This will now throw a "The float ... is not representable as an int, cast occurred" warning.

PHPCSUtils was running into this warning for two tests where a semi-random cache ID was being generated.

Fixed now, by using a slightly smaller range of numbers to pick a random number from, but that should be inconsequential.

Ref:
* https://wiki.php.net/rfc/warnings-php-8-5#casting_out_of_range_floats_to_int